requirements.txt
absl-py==2.2.1
accelerate 
astunparse==1.6.3
Brotli 
certifi 
charset-normalizer
contourpy==1.3.1
cycler==0.12.1
diffusers 
einops==0.8.0
filelock 
flatbuffers==25.2.10
fonttools==4.55.3
fsspec 
gast==0.6.0
gmpy2 
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.12.1
huggingface_hub 
idna 
imageio==2.37.0
importlib_metadata 
Jinja2 
joblib==1.4.2
keras==3.9.1
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe 
matplotlib==3.10.0
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft 
mkl_random 
ml_dtypes==0.5.1
mpmath 
namex==0.0.8
networkx 
numpy 
opencv-python==4.10.0
opencv-python-headless==4.10.0
opt_einsum==3.4.0
optree==0.14.1
packaging 
pillow
protobuf==5.29.3
psutil 
Pygments==2.19.1
pyparsing==3.2.1
PySocks 
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
PyWavelets==1.8.0
PyYAML 
regex 
requests 
rich==14.0.0
safetensors 
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
sewar==0.4.6
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.0
threadpoolctl==3.5.0
tifffile==2025.1.10
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm
triton==3.1.0
typing_extensions 
urllib3 
Werkzeug==3.1.3
wrapt==1.17.2
zipp 
